backend/requirements.txt
fastapi==0.104.1
uvicorn==0.22.0

#new
SQLAlchemy==2.0.13
psycopg2==2.9.6    #linux/mac user use: psycopg2-binary==2.9.6

python-dotenv==1.0.0

alembic==1.13.0

Pydantic[email]

passlib[bcrypt]

pytest==7.4.0
httpx==0.24.1

python-jose==3.3.0

python-multipart==0.0.6

Jinja2==3.1.2
aiofiles==23.1.0